Loadshed middleware, Extendable, Documented
