+++ middleware/loadshed/cpu.go
@@ -1,0 +1,52 @@
+package loadshed
+
+import (
+	"context"
+	"math/rand"
+	"time"
+
+	"github.com/gofiber/fiber/v2/internal/gopsutil/cpu"
+)
+
+// LoadCriteria interface for different types of load metrics.
+type LoadCriteria interface {
+	Metric(ctx context.Context) (float64, error)
+	ShouldShed(metric float64) bool
+}
+
+// CPULoadCriteria for using CPU as a load metric.
+type CPULoadCriteria struct {
+	LowerThreshold float64
+	UpperThreshold float64
+	Interval       time.Duration
+	Getter         CPUPercentGetter
+}
+
+func (c *CPULoadCriteria) Metric(ctx context.Context) (float64, error) {
+	percentages, err := c.Getter.PercentWithContext(ctx, c.Interval, false)
+	if err != nil || len(percentages) == 0 {
+		return 0, err
+	}
+	return percentages[0], nil
+}
+
+func (c *CPULoadCriteria) ShouldShed(metric float64) bool {
+	if metric > c.UpperThreshold*100 {
+		return true
+	} else if metric > c.LowerThreshold*100 {
+		rejectionProbability := (metric - c.LowerThreshold*100) / (c.UpperThreshold - c.LowerThreshold)
+		// #nosec G404
+		return rand.Float64()*100 < rejectionProbability
+	}
+	return false
+}
+
+type CPUPercentGetter interface {
+	PercentWithContext(ctx context.Context, interval time.Duration, percpu bool) ([]float64, error)
+}
+
+type DefaultCPUPercentGetter struct{}
+
+func (r *DefaultCPUPercentGetter) PercentWithContext(ctx context.Context, interval time.Duration, percpu bool) ([]float64, error) {
+	return cpu.PercentWithContext(ctx, interval, percpu)
+}
